Use SemVer to compare local assets to the default ones.
