+++ package.json
@@ -13,8 +13,7 @@
     "coverage": "./node_modules/.bin/nyc report --reporter=text-lcov | coveralls",
     "lint": "./node_modules/.bin/eslint . bin/serverless-artillery",
     "postinstall": "node ./postinstall.js",
-    "publish": "./node_modules/.bin/np",
-    "refresh-lambda-hash": "node ./refresh-lambda-hash.js"
+    "publish": "./node_modules/.bin/np"
   },
   "repository": {
     "type": "git",
